refactor(sse): fold the record resync into SSEClientEffect
SSEClientEffect already owned reconnection semantics - it dispatches SSE_CLIENT_RECONNECTED_EVENT_NAME and resynced the metadata store - so the separate SSERecordsResyncEffect was a second component for the same concern. Both resyncs now live in one debounced listener there.

This also closes the matching gap on the metadata side: resyncMetadataStore was called directly from the graphql-sse connected callback, so it only ran for transport reconnects and never when the keep-alive watchdog or an error tore the stream down and a replacement was built. Driving it off the browser event covers both paths, the same way the record resync does.
