+++ packages/twenty-front/src/modules/sse-db-event/components/SSEClientEffect.tsx
@@ -1,17 +1,22 @@
 import { useHasAccessTokenPair } from '@/auth/hooks/useHasAccessTokenPair';
 import { tokenPairState } from '@/auth/states/tokenPairState';
+import { useListenToBrowserEvent } from '@/browser-event/hooks/useListenToBrowserEvent';
 import { dispatchBrowserEvent } from '@/browser-event/utils/dispatchBrowserEvent';
 import { useResyncMetadataStore } from '@/metadata-store/hooks/useResyncMetadataStore';
+import { useApolloCoreClient } from '@/object-metadata/hooks/useApolloCoreClient';
 import { SSE_CLIENT_RECONNECTED_EVENT_NAME } from '@/sse-db-event/constants/SseClientReconnectedEventName';
+import { SSE_RESYNC_DEBOUNCE_TIME_IN_MS } from '@/sse-db-event/constants/SseResyncDebounceTimeInMs';
 import { useHandleSseClientConnectionRetry } from '@/sse-db-event/hooks/useHandleSseClientConnectionRetry';
 import { activeQueryListenersState } from '@/sse-db-event/states/activeQueryListenersState';
 import { sseClientState } from '@/sse-db-event/states/sseClientState';
 import { useAtomState } from '@/ui/utilities/state/jotai/hooks/useAtomState';
 import { useAtomStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomStateValue';
+import { captureException } from '@sentry/react';
 import { isNonEmptyArray } from '@sniptt/guards';
 import { createClient } from 'graphql-sse';
 import { useCallback, useEffect } from 'react';
 import { isDefined } from 'twenty-shared/utils';
+import { useDebouncedCallback } from 'use-debounce';
 import { REACT_APP_SERVER_BASE_URL } from '~/config';
 import { useStore } from 'jotai';
 
@@ -21,6 +26,30 @@ export const SSEClientEffect = () => {
   const [sseClient, setSseClient] = useAtomState(sseClientState);
   const tokenPair = useAtomStateValue(tokenPairState);
   const { resyncMetadataStore } = useResyncMetadataStore();
+  const apolloCoreClient = useApolloCoreClient();
+
+  const resyncAfterReconnection = useCallback(() => {
+    resyncMetadataStore();
+
+    apolloCoreClient.refetchQueries({ include: 'active' }).catch((error) => {
+      captureException(
+        new Error('Failed to resync records after SSE reconnection', {
+          cause: error,
+        }),
+      );
+    });
+  }, [apolloCoreClient, resyncMetadataStore]);
+
+  const debouncedResyncAfterReconnection = useDebouncedCallback(
+    resyncAfterReconnection,
+    SSE_RESYNC_DEBOUNCE_TIME_IN_MS,
+    { leading: false },
+  );
+
+  useListenToBrowserEvent({
+    eventName: SSE_CLIENT_RECONNECTED_EVENT_NAME,
+    onBrowserEvent: debouncedResyncAfterReconnection,
+  });
 
   const handleSSEClientConnected = useCallback(
     (reconnected: boolean) => {
@@ -33,11 +62,10 @@ export const SSEClientEffect = () => {
       }
 
       if (reconnected) {
-        resyncMetadataStore();
         dispatchBrowserEvent(SSE_CLIENT_RECONNECTED_EVENT_NAME);
       }
     },
-    [store, resyncMetadataStore],
+    [store],
   );
 
   const { handleSseClientConnectionRetry } =
