+++ packages/twenty-front/src/modules/sse-db-event/components/SSEProvider.tsx
@@ -3,7 +3,6 @@ import { SSEClientEffect } from '@/sse-db-event/components/SSEClientEffect';
 import { SSEEventStreamEffect } from '@/sse-db-event/components/SSEEventStreamEffect';
 import { SSEKeepAliveEffect } from '@/sse-db-event/components/SSEKeepAliveEffect';
 import { SSEQuerySubscribeEffect } from '@/sse-db-event/components/SSEQuerySubscribeEffect';
-import { SSERecordsResyncEffect } from '@/sse-db-event/components/SSERecordsResyncEffect';
 import { type ReactNode } from 'react';
 
 type SSEProviderProps = {
@@ -17,7 +16,6 @@ export const SSEProvider = ({ children }: SSEProviderProps) => {
       <SSEEventStreamEffect />
       <SSEQuerySubscribeEffect />
       <SSEKeepAliveEffect />
-      <SSERecordsResyncEffect />
       <MetadataStoreSSEEffect />
       {children}
     </>
