--- packages/twenty-front/src/modules/sse-db-event/constants/SseRecordsResyncDebounceTimeInMs.ts
@@ -1,1 +1,0 @@
-export const SSE_RECORDS_RESYNC_DEBOUNCE_TIME_IN_MS = 1_000;
