+++ packages/twenty-front/src/modules/sse-db-event/constants/SseResyncDebounceTimeInMs.ts
@@ -1,0 +1,1 @@
+export const SSE_RESYNC_DEBOUNCE_TIME_IN_MS = 1_000;
